requirements.txt
requests
tweepy
python-dotenv
pillow
flask
flask-cors
